# src/main/resources/openapi/action.yml
openapi: 3.0.3
info:
  title: Temporal Rift — Action API
  description: Action round submission endpoints
  version: 1.0.0

servers:
  - url: /api/v1

security:
  - bearerAuth: [ ]

tags:
  - name: Action

paths:
  /games/{gameId}/eras/{eraNumber}/rounds/{roundNumber}/actions:
    post:
      tags: [ Action ]
      operationId: submitAction
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eraNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: roundNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitActionRequest'
      responses:
        '202':
          description: Action submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitActionResponse'
        '409':
          description: |
            Round already closed (`code: 409-01`) or player already submitted (`code: 409-02`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '422':
          description: |
            Validation error:
            - Card not in hand (`code: 422-01`)
            - Player jammed (`code: 422-02`)
            - Invalid target (`code: 422-03`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /games/{gameId}/eras/{eraNumber}/rounds/{roundNumber}/status:
    get:
      tags: [ Action ]
      operationId: getRoundStatus
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eraNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: roundNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Round submission status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundStatusResponse'
        '404':
          description: |
            Game or round not found (`code: 404-01`). Also returned when the requesting player is
            not a participant of the game, so game and round existence is never disclosed to
            outsiders.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ActionType:
      type: string
      enum: [ CARD, SPECIAL ]

    SpecialAction:
      type: string
      enum:
        - ANNIHILATE
        - CORRUPT
        - CASCADE
        - FORESIGHT
        - SEAL
        - FULFILLMENT
        - REWRITE
        - MIMIC
        - OBSCURE
        - THREAD
        - TAPESTRY
        - UNRAVEL
        - RALLY
        - EXPOSE
        - MOMENTUM

    ActionSubmissionStatus:
      type: string
      enum: [ SUBMITTED ]

    RoundStatus:
      type: string
      enum: [ OPEN, CLOSED ]

    SubmitActionRequest:
      type: object
      required: [ actionType ]
      properties:
        actionType:
          $ref: '#/components/schemas/ActionType'
      discriminator:
        propertyName: actionType
        mapping:
          CARD: '#/components/schemas/CardActionRequest'
          SPECIAL: '#/components/schemas/SpecialActionRequest'

    CardActionRequest:
      allOf:
        - $ref: '#/components/schemas/SubmitActionRequest'
        - type: object
          required: [ cardInstanceId, targetEventId, targetOutcomeId ]
          properties:
            cardInstanceId:
              type: string
              format: uuid
            targetEventId:
              type: string
              format: uuid
            sourceOutcomeId:
              type: string
              format: uuid
              description: Required when the submitted card resolves as SWING and must differ from targetOutcomeId.
            targetOutcomeId:
              type: string
              format: uuid
              description: Required for card submissions and, for SWING, must differ from sourceOutcomeId.

    SpecialActionRequest:
      allOf:
        - $ref: '#/components/schemas/SubmitActionRequest'
        - type: object
          required: [ specialAction, targetEventId, targetOutcomeId ]
          properties:
            specialAction:
              $ref: '#/components/schemas/SpecialAction'
            targetEventId:
              type: string
              format: uuid
            targetOutcomeId:
              type: string
              format: uuid
            targetPlayerId:
              type: string
              format: uuid

    SubmitActionResponse:
      type: object
      required: [ gameId, eraNumber, roundNumber, playerId, status, roundClosed ]
      properties:
        gameId:
          type: string
          format: uuid
        eraNumber:
          type: integer
          minimum: 1
        roundNumber:
          type: integer
          minimum: 1
        playerId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ActionSubmissionStatus'
        roundClosed:
          type: boolean

    RoundStatusResponse:
      type: object
      required: [ eraNumber, roundNumber, status, timerRemainingSeconds, submittedCount, totalPlayers, pendingPlayerIds ]
      properties:
        eraNumber:
          type: integer
          minimum: 1
        roundNumber:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/RoundStatus'
        timerRemainingSeconds:
          type: integer
          minimum: 0
        submittedCount:
          type: integer
        totalPlayers:
          type: integer
        pendingPlayerIds:
          type: array
          items:
            type: string
            format: uuid

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type.
          format: uri
          maxLength: 1024
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          format: int32
          minimum: 100
          maximum: 599
        title:
          type: string
          description: A short, human-readable summary of the problem type. It should not change from occurrence to occurrence of the problem, except for purposes of localization.
          maxLength: 1024
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          maxLength: 4096
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.
          maxLength: 1024
        code:
          type: string
          description: An API specific error code aiding the provider team understand the error based on their own potential taxonomy or registry.
          maxLength: 50
        errors:
          type: array
          description: An array of error details to accompany a problem details response.
          maxItems: 1000
          items:
            type: object
            description: An object to provide explicit details on a problem towards an API consumer.
            properties:
              detail:
                type: string
                description: A granular description on the specific error related to a body property, query parameter, path parameters, and/or header.
                maxLength: 4096
              pointer:
                type: string
                description: A JSON Pointer to a specific request body property that is the source of error.
                maxLength: 1024
              parameter:
                type: string
                description: The name of the query or path parameter that is the source of error.
                maxLength: 1024
              header:
                type: string
                description: The name of the header that is the source of error.
                maxLength: 1024
              code:
                type: string
                description: A string containing additional provider specific codes to identify the error context.
                maxLength: 50
            required:
              - detail
      required:
        - detail